Backend/requirements.txt
alembic==1.15.2
annotated-types==0.7.0
anyio==4.4.0
async-timeout==4.0.3
asyncpg==0.30.0
Authlib==1.6.0
bcrypt==4.3.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.2
click==8.1.7
colorama==0.4.6
cryptography==44.0.3
dnspython==2.7.0
ecdsa==0.19.1
email_validator==2.2.0
exceptiongroup==1.2.2
fastapi==0.112.2
fastapi-cli==0.0.7
greenlet==3.2.0
h11==0.14.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.8
httptools==0.6.1
httpx==0.28.1
Hypercorn==0.17.3
hyperframe==6.1.0
idna==3.8
itsdangerous==2.2.0
Jinja2==3.1.6
jose==1.0.0
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.25
langchain-core==0.3.65
langchain-text-splitters==0.3.8
langsmith==0.3.45
Mako==1.3.10
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
orjson==3.10.16
packaging==24.2
pillow==11.2.1
priority==2.0.0
psycopg2==2.9.10
psycopg2-binary==2.9.9
pyasn1==0.4.8
pycparser==2.22
pydantic==2.8.2
pydantic-extra-types==2.10.3
pydantic-settings==2.4.0
pydantic_core==2.20.1
Pygments==2.19.1
PyPDF2==3.0.1
pytesseract==0.3.13
python-dotenv==1.0.1
python-jose==3.4.0
python-multipart==0.0.20
PyYAML==6.0.2
requests==2.32.4
requests-toolbelt==1.0.0
rich==14.0.0
rich-toolkit==0.14.1
rsa==4.9.1
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.40
SQLAlchemy-Utils==0.41.2
sqlmodel==0.0.24
starlette==0.38.4
taskgroup==0.2.2
tenacity==9.1.2
tomli==2.2.1
typer==0.15.2
typing_extensions==4.12.2
ujson==5.10.0
urllib3==2.4.0
uvicorn==0.30.6
uvloop==0.21.0
watchfiles==0.24.0
websockets==13.0.1
wsproto==1.2.0
zstandard==0.23.0
